requirements.txt
meshtastic
pypubsub
#sqlite3
